Update domain unit tests
